relocate CreditId class file
